# Conflicts:
#	src/main/java/com/fawryEx/storyEx/controller/StockController.java
#	src/main/java/com/fawryEx/storyEx/entity/Stock.java
#	src/main/java/com/fawryEx/storyEx/repository/StockRepository.java
#	src/main/java/com/fawryEx/storyEx/service/StockService.java

+++ src/main/java/com/fawryEx/storyEx/service/StoreService.java
@@ -43,7 +43,11 @@ public class StoreService {
                 .orElseThrow(() -> new StoreNotFoundException("Store not found"));
     }
 
+    @Transactional
     public void deleteStore(Long id) {
+        if (!storeRepository.existsById(id)) {
+            throw new StoreNotFoundException("the id = " + id + " not found");
+        }
         storeRepository.deleteById(id);
     }
